requirements.txt
pandas==1.3.3
s3fs==2021.10.1
apache-airflow==2.7.0